owen add(api/routes/jobs) create de jobs routes
