+++ api/package.json
@@ -21,10 +21,11 @@
     "dotenv": "^8.2.0",
     "express": "^4.17.1",
     "morgan": "^1.10.0",
+    "nodemailer": "^6.7.8",
+    "nodemon": "^2.0.7",
     "pg": "^8.5.1",
     "sequelize": "^6.3.5",
-    "supertest-as-promised": "^4.0.2",
-    "nodemon": "^2.0.7"
+    "supertest-as-promised": "^4.0.2"
   },
   "devDependencies": {
     "chai": "^4.2.0",
