+++ api/src/routes/index.js
@@ -1,12 +1,14 @@
 const { Router } = require('express');
-const users = require("./users/users.routes")
+const users = require("./users/users.route")
 const verify = require("./verify/verify.route")
+const jobs = require("./jobs/jobs.route")
 // Importar todos los routers;
 // Ejemplo: const authRouter = require('./auth.js');
 const router = Router();
 
 router.use("/users", users);
-router.use("/verify", verify)
+router.use("/verify", verify);
+router.use("/jobs", jobs);
 // router.use("/x", x)
 
 // Configurar los routers
