+++ api/src/routes/users/users.route.js
@@ -66,30 +66,6 @@ users.post("/", async (req, res, next) =>{
     
 })
 
-// RUTA QUE CREA TRABAJOS
-users.post("/createJobs", async (req, res, next) =>{
-    const { name } = req.body;
-    const jobsMinuscule = name.toLowerCase();
-    try {
-        const [newJob, created] = await Profession.findOrCreate({
-            where:{
-                name: jobsMinuscule,
-            },
-            defaults:{
-                name: jobsMinuscule,
-            }
-        })
-
-        if(!created)  res.status(200).send(`The Profession cannot be created, the Job "${jobsMinuscule}" has already exist`);
-        return res.status(201).send(`The Profession "${jobsMinuscule}" created successfully`);
-        
-    } catch (error) {
-        console.log(error)
-        next(error)
-    }
-    
-})
-
 // RUTA QUE CREA RESEÑAS
